+++ changelog.md
@@ -4,3 +4,4 @@
 - Pressing `view workout` won't produce duplicate results anymore
 - Added next exercise info to rest screen
 - Changed location of default workout to user folder
+- Added the ability to create a new workout file